Fix the "requires" decorator
