+++ weasyprint/text.py
@@ -117,6 +117,8 @@ ffi.cdef('''
         int height;
     } PangoRectangle;
 
+    int pango_version (void);
+
     double pango_units_to_double (int i);
     int pango_units_from_double (double d);
     void g_object_unref (gpointer object);
